\name{FastKM-package}
\alias{FastKM-package}
\alias{FastKM}
\docType{package}
\title{
A Fast Multiple-Kernel Method Based on a Low-Rank Approximation
}
\description{
A computationally efficient and statistically rigorous fast Kernel Machine method for multi-kernel analysis. The approach is based on a low-rank approximation to the nuisance effect kernel matrices. The algorithm is applicable to continuous, dichotomous, and survival traits and is implemented using the existing single-kernel analysis software 'SKAT' or 'coxKM'.}
\details{
\tabular{ll}{
Package: \tab FastKM\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2022-06-05\cr
License: \tab GPL-2\cr
}
}
\author{
Shannon T. Holloway, Rachel Marceau, Wenbin Lu, Michele M. Sale, Bradford B. Worrall, Stephen R. Williams, Fang-Chi Hsu, and Jung-Ying Tzeng

Maintainer: Shannon T. Holloway <shannon.t.holloway@gmail.com>
}
\references{
Marceau, R., Lu, W., Holloway, S. T., Sale, M. M., Worrall, B. B., Williams, S. R., Hsu, F-C., and Tzeng, J-Y. (2015). 
A Fast Multiple-Kernel Method with Applications to Detect Gene-Environment Interaction. Genetic Epidemiology, 39, 456-468.
}

